services/backend/requirements.txt
aniso8601==9.0.1
attrs==22.1.0
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
python-dotenv==0.21.0
PyJWT==2.6.0
bcrypt==3.2.0
Werkzeug==2.1.1
psycopg2-binary==2.9.3
